Fix mixed content errors for loadDemo (#4)
* Automatically fix mixed content.

On some versions and security settings of chrome, the loadDemo function will fail when mixed content is loaded.

As an example, going to http://br.hogclangaming.com/pr1/ and then right clicking a link and copying the link address might return http://br.hogclangaming.com/pr1/auto_2019_10_31_23_35_26.bf2demo. However, because this is using http and not the more secure https, it will cause chrome to automatically block it. My fix solves this by using a URL object to change the protocol to whatever the window's protocol is. This guarantees that they will always match. Which makes testing locally or in production easier.

* Spelling

Fixed bad spelling